Fix ambiguity with virtual inheritance
Virtually inherit the drawable class in both ISprite and SpriteWrapper
and add a new Sprite class member to SpriteWrapper. Override the virtual
draw function from drawable to point the target to the sprite class member
so that the laser cannon can be drawn onto the window.

This will also allow the tests to remain unchanged and still pass.