Create move for laser cannon
Pass the first test for the move function by adding the float passed in
to the x position of the laser cannon. This only needs to be applied to
the x position and not the y position because the player will always stay
in the same vertical position in the game.