Remove sound and sprite wrapper implementation
Reset functionality to not use the sound and sprite wrappers. If tests
are compiled they will fail so these will be updated soon.

Wrappers will need to be redone so that the set up for each object can be
done in their constructors, rather than having to create all the sounds
and sprites before instantiating the Laser Cannon, Laser and Invaders.