Add return statement to break out of loops
Optimise the efficiency of the function by returning once a valid collision
has been detected as the laser cannot be in two places at once. i.e. if the
laser is colliding with one invader, it can't be colliding with any others
since they are all in different positions. Therefore, it would be meaningless
to continue the nested for loops to check collisions with the other invaders.