Reset invader positions
Call setPosition at level up with the new level up position. This has the
same x component as the invader's original position but a y position shifted
down from the invader's original position that increases as the level
increases. This is so that the difficulty of each level gradually increases.