Reorganise invader classes
Use the protected keyword to allow the invader subclasses to access the
width, height, and points members from the Invader class. This avoids
repeating the same code with the invader getWidth, getHeight, and getPoints
function.

This method will also be implemented across all classes that share the
same functions from ObjectInterface.