Reorganise class inheritance
Since every object on the board has some of the their functions and members
the same, give the ObjectInterface class these members to avoid repeating
code throughout the classes. This only really applies to the constructors
and setting values at instantiation. This is because the i abstract classes
for each object still need to have virtual functions that can be mocked in
the tests. This change will slightly tidy up the class inheritance tree
though.