Add virtual functions to common inheritance
Add some of the the virtual functions from the i abstract classes to
ObjectInterface as they are common amongst all of them. Add these to
the necessary mock classes since the virtual functions will now be
inherited by the mock classes and will need to be mocked in order for
the mock classes to be instantiated.