Make resetPosition
Create a new function that will deal with solely resetting the laser cannon
position back to its starting point. This will be need to differentiate
between resetting the cannon after losing a life and after game over.

This is because after game over, the lives will need to be reset as well.