Update sprite position when moving
After altering the position class member, update the sprite position
so that the sprite moves along with the object.

Also, rather than saving the pointers to sprites on the heap in tests,
save them on the stack for faster memory allocation, avoiding leaks, and
general better coding practice. The direction class member was actually
not initialised in the constructor, meaning the tests for the direction
actually shouldn't have passed. However, for some reason, saving the
sprite pointers on the heap caused them to pass, and therefore
invalidated some of the tests. This will now be fixed along with initalising
the direction member to 1 in the constructor.