Restart spawn function
Rewrite the first test to check it sets the direction to 1. This will be
needed since when Ridley reaches one end of the board, the direction will
need to be reversed so that it can start moving back the other way. This
will be done in the spawn function before it starts moving.

Therefore, initial starting value of direction needs to be set to -1 and
also its reset value in the reset function.