Fill space on enlarging window
Using a conditional, check which dimension of the game can fill the window
and adjust the other dimension accordingly to retain the aspect ratio.

Add this whole resize functionality in a conditional of its own that will
be applied only when both the width and height dimensions are both bigger
than their original values.