Adjust view for any window resize
After trying to make a solution for diminishing the window, it turns out
this solution also successfully adjusts the view for enlarging the window
as well.

Start by finding which dimension of the game can completely fill the
corresponding dimension of the window. Then adjust the blank space around
the other dimension to maintain the aspect ratio of 1536:1344.

Do this with use of a FloatRect to set the window's view. The origin
values are the width or height of the blank space added around the game.
The newWidth and newHeight values are the game window dimensions added to
the blank spaces.

Both the origin values and newHeight/Width values use the in-game
coordinates.